Legacy types moved to restoreUtils file.
