--- src/legacyTypes.ts
@@ -1,6 +1,0 @@
-import type { BaseRenameItem } from "./types.js";
-
-export interface LegacyRenameItem extends BaseRenameItem {
-  sourcePath: string;
-}
-export type LegacyRenameList = LegacyRenameItem[];
